+++ tests/TestCase.php
@@ -48,16 +48,32 @@ abstract class TestCase extends TestkitTest_Case {
 		$client->set( $key3, $val3, $grp2 );
 
 		// Get key1, delete it, and assert it's gone.
-		$this->assertSame( $val1, $client->get( $key1, $grp1 ) );
+		$this->assertSame( $val1, $client->value( $key1, $grp1 ) );
 		$client->delete( $key1, $grp1 );
-		$this->assertNull( $client->get( $key1, $grp1 ) );
+		$this->assertNull( $client->value( $key1, $grp1 ) );
 
 		// Get key2, flush group1, and assert it's gone.
-		$this->assertSame( $val2, $client->get( $key2, $grp1 ) );
+		$this->assertSame( $val2, $client->value( $key2, $grp1 ) );
 		$client->flush_group( $grp1 );
-		$this->assertNull( $client->get( $key2, $grp1 ) );
+		$this->assertNull( $client->value( $key2, $grp1 ) );
 
 		// key3 should still be there.
-		$this->assertSame( $val3, $client->get( $key3, $grp2 ) );
+		$this->assertSame( $val3, $client->value( $key3, $grp2 ) );
+
+		// Put key1 and key2 back in.
+		$client->set( $key1, $val1, $grp1 );
+		$client->set( $key2, $val2, $grp1 );
+
+		// The values in the group should be key1 or key2.
+		foreach ( $client->group( $grp1 ) as $item ) {
+			$this->assertTrue( $item->value === $val1 || $item->value === $val2 );
+
+			// Delete the item.
+			$item->delete();
+		}
+
+		// Assert that the group is empty again.
+		$this->assertNull( $client->value( $key1, $grp1 ) );
+		$this->assertNull( $client->value( $key2, $grp1 ) );
 	}
 }