+++ tests/Unit/InMemoryCacheTest.php
@@ -25,38 +25,38 @@ class InMemoryCacheTest extends TestCase {
 
 		// Two fetches of the same key should only result in one query.
 		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key1', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
 		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
 
 		$big_pit->set( 'key1', 'value1', 'group1' );
 
 		// Value has changed, so there should be another query.
 		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
 		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
 
 		// Fetching it again should not result in another query.
-		$big_pit->get( 'key1', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
 		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
 
 		$big_pit->delete( 'key1', 'group1' );
 
 		// Value has changed, so there should be another query.
 		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
 		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
 
 		$big_pit->set( 'key1', 'value1', 'group1' );
 		$big_pit->set( 'key2', 'value2', 'group1' );
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key2', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
+		$big_pit->value( 'key2', 'group1' );
 		$big_pit->flush_group( 'group1' );
 
 		// All values have changed, so there should be queries for each value in the group that was set.
 		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key2', 'group1' );
+		$big_pit->value( 'key1', 'group1' );
+		$big_pit->value( 'key2', 'group1' );
 		$this->assertSame( 2, $wpdb->num_queries - $num_queries_before );
 	}
 
@@ -67,6 +67,6 @@ class InMemoryCacheTest extends TestCase {
 		$big_pit = new Big_Pit();
 		$big_pit->boot();
 		$big_pit->set( 'key1', (object) [ 'foo' => 'bar' ], 'group1' );
-		$this->assertNotSame( $big_pit->get( 'key1', 'group1' ), $big_pit->get( 'key1', 'group1' ) );
+		$this->assertNotSame( $big_pit->value( 'key1', 'group1' ), $big_pit->value( 'key1', 'group1' ) );
 	}
 }